Refactor to move out builder
