+++ app/frontend-webapp/auth/policies.ts
@@ -1,37 +1,5 @@
+import { AuthorizationPolicyBuilder } from "@/auth/AuthorizationPolicyBuilder";
 import { permissions } from "@/auth/permissions";
-
-type AuthorizationPolicy = {
-  isAuthorized: (userPermissions: string[]) => boolean;
-  getPermissions: () => string[];
-};
-
-class AuthorizationPolicyBuilder {
-  private policies: AuthorizationPolicy[] = [];
-
-  Combine(...policies: AuthorizationPolicy[]): AuthorizationPolicyBuilder {
-    this.policies.push(...policies);
-    return this;
-  }
-
-  RequireRealmRoles(...roles: string[]): AuthorizationPolicyBuilder {
-    this.policies.push({
-      isAuthorized: (userPermissions: string[]) =>
-        roles.every((role) => userPermissions.includes(role)),
-      getPermissions: () => roles,
-    });
-    return this;
-  }
-
-  Build(): AuthorizationPolicy {
-    const combinedPolicy: AuthorizationPolicy = {
-      isAuthorized: (userPermissions: string[]) =>
-        this.policies.every((policy) => policy.isAuthorized(userPermissions)),
-      getPermissions: () =>
-        this.policies.flatMap((policy) => policy.getPermissions()),
-    };
-    return combinedPolicy;
-  }
-}
 
 export class AuthorizationPolicies {
   // TODO: check if the user has a token at all
@@ -40,12 +8,12 @@ export class AuthorizationPolicies {
 
   public static readonly CanViewSiteReports = new AuthorizationPolicyBuilder()
     .Combine(AuthorizationPolicies.IsAuthenticatedUser)
-    .RequireRealmRoles(permissions.ViewSiteReports)
+    .RequireRoles(permissions.ViewSiteReports)
     .Build();
 
   public static readonly CanGenerateSyntheticData =
     new AuthorizationPolicyBuilder()
       .Combine(AuthorizationPolicies.IsAuthenticatedUser)
-      .RequireRealmRoles(permissions.GenerateSyntheticData)
+      .RequireRoles(permissions.GenerateSyntheticData)
       .Build();
 }
